// src/types/enum/channel.enum.ts
export enum ROLE_TYPES {
    CAPTAIN = 'captain',
    SUB_CAPTAIN = 'sub_captain',
    MEMBER = 'member',
}

export enum MESSAGE_TYPES {
    TEXT = 'text',
    IMAGE = 'image',
    VIDEO = 'video',
    FILE = 'file',
    AUDIO = 'audio',
    EMOJI = 'emoji',
    SYSTEM = 'system',
    OTHER = 'other'
}